+++ CHANGELOG.md
@@ -2,7 +2,8 @@
 
 ### ⚠️ 非兼容性变更
 
-* [^] `Progress` 组件的 `state` prop 更名为 `status`。
+* [^] `Progress` 组件的 `state` prop 更名为 `status`。`state` 将在 `1.0.0` 移除。
+* [^] `Schedule` 组件的 `shortcuts-display` prop 值 `expand`/`collapse` 分别更名为 `inline`/`popup`。旧的值将在 `1.0.0` 移除。
 
 ### 💡 主要变更